fix(mcp): close two escrow.test.ts bugs deferred from #277
Both findings from the 2026-05-14 audit's pr-test-analyzer:

  1. Three tests in `deposit_escrow SOLVELA_ESCROW_MODE gate` asserted
     `process.env['SOLVELA_ESCROW_MODE'] === 'enabled'` — JS string
     equality, not any SDK code. They would have passed even if
     index.ts deleted its escrow gating entirely. Removed; the
     `getTools()` block below already covers the real production
     path.

  2. `deposit_escrow strict amount parsing (H3)` re-implemented the
     parse regex inline as a test-local helper. A production regex
     change would not have failed the tests. Extracted the real
     function to `src/parse-amount.ts` and have the test import it.

Now the parseStrictUsdc tests run against the real function — a
regression in the index.ts amount validation will actually fail them.

66 tests pass (down from 69 because three dead tests were removed).
